+++ satan/dl-satan.el
@@ -17,6 +17,7 @@
 (require 'dl-satan-tools-hippocampus)
 (require 'dl-satan-tools-inbox)
 (require 'dl-satan-tools-agenda)
+(require 'dl-satan-tools-activity)
 (require 'dl-satan-mode)
 (require 'dl-satan-context)
 (require 'dl-satan-output)